+++ view/utilisateur/detail.php
@@ -6,8 +6,8 @@
     Adresse de livraison : '.$uadresseL.' <br>
     Identifiant de carte bleu : '.$uidCB.' <br> 
     Email : '.$uemail.'</p>';
-    //if (Session::is_user($ulogin) || Session::is_admin()) {
+    if (Session::is_user($ulogin) || Session::is_admin()) {
         echo ('<a href="index.php?controller=utilisateur&action=delete&loginUtilisateur=' . rawurlencode($ulogin) . '" style="color:black"> <br> <p> Cliquez ici pour supprimer cette utilisateur ! </p> </a>');
         echo ('<a href="index.php?controller=utilisateur&action=update&loginUtilisateur=' . rawurlencode($ulogin) . '" style="color:black"> <p> Cliquez ici pour modifier cette utilisateur ! </p> </a>');
-    //}
+    }
 ?>
